requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
branca==0.7.2
Brotli==1.1.0
brotlicffi==1.1.0.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dotmap==1.3.30
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.20.0
fastparquet==2024.5.0
feather-format==0.4.1
filelock==3.15.4
folium==0.17.0
fonttools==4.53.0
formulaic==1.0.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gcsfs==2024.6.1
GDAL==3.9.1
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-api-core==2.19.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.1.2
google-resumable-media==2.7.0
googleapis-common-protos==1.63.2
graphlib-backport==1.0.3
grpcio==1.62.2
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
hydra_zen==0.13.0
hyperframe==6.0.1
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflate64==1.0.0
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==8.6.2
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lazy_loader==0.4
lifelines==0.29.0
lightning-utilities==0.11.3.post0
line-profiler==4.1.1
llvmlite==0.43.0
mapclassify==2.6.1
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multivolumefile==0.2.3
munkres==1.1.4
mypy==1.10.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.8.7
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.4
osqp==0.6.7.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.4
pathtools==0.1.2
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.0
pkgconfig==1.5.5
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.21.1
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pycox==0.2.3
pycparser==2.22
pycryptodomex==3.19.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyogrio==0.9.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
pyppmd==1.1.0
pyproj==3.6.1
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyu2f==0.1.5
pyvips==2.2.3
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
pyzstd==0.16.0
qdldl==0.1.5.post2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
rsa==4.9
ruff==0.5.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.7.1
setproctitle==1.3.3
setuptools==70.1.1
shapely==2.0.4
sip==6.7.12
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.2
sympy==1.12.1
tenacity==8.4.2
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
torch==2.2.2
torch_cluster==1.6.3+pt22cu121
torch_geometric==2.5.2
torch_scatter==2.1.2+pt22cu121
torch_sparse==0.6.18+pt22cu121
torchmetrics==1.4.0.post0
torchsurv==0.1.2
torchtuples==0.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
typing-utils==0.1.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wandb==0.16.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xyzservices==2024.6.0
yarl==1.9.4
zipfile-deflate64==0.0.0
zipp==3.19.2
zstandard==0.22.0
